requirements.txt
altair==4.2.0
apispec==5.2.2
arrow==1.2.2
attrs==19.3.0
Automat==0.8.0
aws-lambda-builders==1.18.0
aws-sam-cli==1.55.0
aws-sam-translator==1.50.0
awscli==1.25.60
awscli-local==0.20
backports.zoneinfo==0.2.1
binaryornot==0.4.4
blinker==1.4
boto3==1.24.59
botocore==1.27.59
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
chevron==0.14.0
click==7.1.2
cloud-init==22.1
colorama==0.4.3
colormap==1.0.4
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.5
cookiecutter==2.1.1
cryptography==37.0.4
cycler==0.11.0
dateparser==1.1.1
dbus-python==1.2.16
decorator==5.1.1
dill==0.3.2
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dnslib==0.9.20
dnspython==2.2.1
docker==4.2.2
docutils==0.16
ecdsa==0.18.0
entrypoints==0.3
filelock==3.8.0
Flask==1.1.4
fonttools==4.37.4
gitdb==4.0.10
GitPython==3.1.30
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==3.0.0
jinja2-time==0.2.0
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
localstack==1.0.4
localstack-client==1.36
localstack-ext==1.0.4
MarkupSafe==2.0.1
matplotlib==3.6.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.2
oauthlib==3.1.0
opencv-python==4.7.0.68
packaging==21.3
pandas==1.5.2
pbr==5.10.0
pexpect==4.6.0
Pillow==9.2.0
platformdirs==2.5.2
plux==1.3.1
prettytable==3.5.0
protobuf==3.20.3
psutil==5.9.1
pyaes==1.6.1
pyarrow==10.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydeck==0.8.0
pyecharts==2.0.0
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-jose==3.3.0
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==5.3.1
regex==2021.9.30
requests==2.25.1
requests-unixsocket==0.2.0
rich==12.5.1
rsa==4.7.2
s3transfer==0.6.0
SecretStorage==2.3.1
semver==2.13.0
serverlessrepo==0.1.10
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.3
ssh-import-id==5.10
stevedore==4.0.0
streamlit==1.16.0
streamlit-drawable-canvas==0.9.2
streamlit-echarts==0.4.0
streamlit-option-menu==0.3.2
systemd-python==234
tabulate==0.8.10
tailer==0.4.1
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.2
toolz==0.12.0
tornado==6.2
Twisted==18.9.0
typing_extensions==4.3.0
tzlocal==3.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
validators==0.20.0
virtualenv==20.16.6
wadllib==1.3.3
watchdog==2.1.2
wcwidth==0.2.5
websocket-client==1.4.0
Werkzeug==1.0.1
zipp==1.0.0
zope.interface==4.7.1
